src/sage/libs/giac/__init__.py: delete a disabled doctest
This file has a test that computes a Groebner basis using multiple
threads, but it is currently skipped. I have tried to de-skip it, but
then giac segfaults while trying to read past the end of a vector. The
author is aware of this problem and refuses to fix it, so I am
deleting the no-op rather than figure out what is necessary to
re-enable it.