src/sage/libs/giac/giac.pyx: add more Pygen() wrappers in redim()
In redim(), we subtract the current dimensions of the matrix from the
two size arguments. The arguments are "integer like," but the matrix
dimensions are Pygens, so this can fail pretty easily if you e.g.
pass in a python int for the new size: python ints don't know how
to subtract a Pygen from themselves.

This is easily fixed by wrapping the arguments in Pygen().