*.*: rename the module and convert it to a namespace package
Testing this package was difficult for two reasons. First, it lived in
the sage.libs namespace, and sage.libs is currently a non-namespace
package. That meant that you had to *install* sagemath-giac before you
could test it; otherwise, "from sage.libs..."  would look (only) in
the sage hierarchy, and not your build directory -- even if PYTHONPATH
was set.

Second, the __init__.py in sage/libs/giac caused the exact same
problem with our own build and source directories after building the
project with meson. Any "from sage.libs.giac.giac..." import would
fail because the presence of __init__.py in the source directory would
hide the shared library in the build directory.

To address these issues, we overhaul the naming scheme:

  * sage.libs.giac is renamed to sagemath_giac (this does not have to
    be permanent, if namespace packages in cython are ever fixed). It
    also does not have to affect most users, because sage can
    re-export the contents of this package under sage.libs.giac.

  * __init__.py is split into gb.py (for groebner_basis) and
    context.py (for the local_settings context manager / decorator).

Afterwards it is possible to test the package without installing it,
directly from the build directory.