src/sagemath_giac/giac.pyx: drop matrix latex() example
In the current version of Sage, calling latex(M) on a matrix M will
launch a LaTeX process in a temporary directory to determine if a
usable LaTeX installation is present. This is irrelevant to what we
are doing, and it happens to fail sometimes, so I think we can do
without it. There's still one latex() test that does not use a
matrix.