meson.build: search PATH for "python"
The default find_installation() gets its "python" either from meson or
from meson-python. When those are the system copies, we can mis-detect
the system copy of python rather than the venv python, if we happen to
be in a venv.

To fix it, we pass 'python' to find_installation() explicitly.