src/sagemath_giac/meson.build: improved cython include-dir hack
Instead of looping through sys.path to construct the cython include
dirs, we now use python to find the directory that sage lives in, and
pass that to cython. This should work in the same scenarios that the
existing hack was created for while additionally supporting --editable
installs of sage.